requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asyncio==3.4.3
attrs==23.2.0
azure-core==1.30.1
azure-cosmos==4.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
typing_extensions==4.12.0
urllib3==2.2.1
yarl==1.9.4
